[refactor] :recycle: move functions from models to services
