+++ src/controllers/nftController.ts
@@ -9,8 +9,8 @@ import { getDynamicGas } from '../utils/dynamicGas';
 import { SIGNING_KEY } from '../constants/environment';
 import NFTModel, { INFT, INFTMetadata } from '../models/nft';
 import { getNetworkConfig } from '../services/networkService';
-import { IUser, getWalletByPhoneNumber } from '../models/user';
 import { createUserWithWallet } from '../services/userService';
+import { getWalletByPhoneNumber } from '../services/walletService';
 import { sendMintNotification } from '../services/notificationService';
 import { defaultNftImage, networkChainIds } from '../constants/blockchain';
 import { returnErrorResponse, returnSuccessResponse } from '../utils/responseFormatter';
