+++ src/models/user.ts
@@ -33,18 +33,3 @@ const userSchema = new Schema<IUser>({
 });
 
 export const User = model<IUser>('User', userSchema, 'users');
-
-/**
- * Función para obtener el wallet basado en el número de teléfono
- * @param {string} phoneNumber - El número de teléfono a buscar
- * @returns {Promise<string | null>} La dirección del wallet o null si no se encuentra
- */
-export async function getWalletByPhoneNumber(phoneNumber: string): Promise<string | null> {
-  try {
-    const user = await User.findOne({ phone_number: phoneNumber }).select('wallet').exec();
-    return user ? user.wallet : null; // Retorna la wallet si se encuentra, de lo contrario null
-  } catch (error) {
-    console.error('Error al obtener la wallet:', error);
-    throw new Error('No se pudo obtener la wallet');
-  }
-}
