+++ src/services/walletService.ts
@@ -332,3 +332,18 @@ export async function getTokenInfo(tokens: IToken[], chanId: number): Promise<To
     rateUSD: prices.get(token.symbol) || 0
   }));
 }
+
+/**
+ * Función para obtener el wallet basado en el número de teléfono
+ * @param {string} phoneNumber - El número de teléfono a buscar
+ * @returns {Promise<string | null>} La dirección del wallet o null si no se encuentra
+ */
+export async function getWalletByPhoneNumber(phoneNumber: string): Promise<string | null> {
+  try {
+    const user = await User.findOne({ phone_number: phoneNumber }).select('wallet').exec();
+    return user ? user.wallet : null; // Retorna la wallet si se encuentra, de lo contrario null
+  } catch (error) {
+    Logger.error('Error al obtener la wallet:', error);
+    throw new Error('No se pudo obtener la wallet');
+  }
+}
